feat: add exchange lookup
New dictionary request implemented that responds with a lot of static info, including a list of exchanges. For now only this list is parsed and stored, and can be used to get the name and other info of the exchange based on id.

Also added a new file with helper functions for common operations like lookups in the context.